Extract out 'net/storage_thread' into separate components.
The logic for `storage_thread.rs` (in `net` and `net_traits`) is
already isolated from the rest of the other modules in the crates.
