+++ components/script/dom/bindings/trace.rs
@@ -60,7 +60,6 @@ use msg::constellation_msg::{PipelineId, SubpageId, WindowSizeData, WindowSizeTy
 use net_traits::image::base::{Image, ImageMetadata};
 use net_traits::image_cache_thread::{ImageCacheChan, ImageCacheThread};
 use net_traits::response::HttpsState;
-use net_traits::storage_thread::StorageType;
 use net_traits::{Metadata, NetworkError};
 use offscreen_gl_context::GLLimits;
 use profile_traits::mem::ProfilerChan as MemProfilerChan;
@@ -82,6 +81,7 @@ use std::rc::Rc;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, AtomicUsize};
 use std::sync::mpsc::{Receiver, Sender};
+use storage_traits::storage_thread::StorageType;
 use string_cache::{Atom, Namespace, QualName};
 use style::attr::{AttrIdentifier, AttrValue};
 use style::element_state::*;
