+++ components/script/dom/storage.rs
@@ -14,10 +14,10 @@ use dom::event::{Event, EventBubbles, EventCancelable};
 use dom::storageevent::StorageEvent;
 use dom::urlhelper::UrlHelper;
 use ipc_channel::ipc;
-use net_traits::storage_thread::{StorageThread, StorageThreadMsg, StorageType};
 use page::IterablePage;
 use script_runtime::ScriptChan;
 use script_thread::{MainThreadRunnable, MainThreadScriptChan, ScriptThread};
+use storage_traits::storage_thread::{StorageThread, StorageThreadMsg, StorageType};
 use task_source::dom_manipulation::DOMManipulationTask;
 use url::Url;
 use util::str::DOMString;
