+++ components/script/dom/window.rs
@@ -48,7 +48,6 @@ use msg::constellation_msg::{WindowSizeData, WindowSizeType};
 use msg::webdriver_msg::{WebDriverJSError, WebDriverJSResult};
 use net_traits::ResourceThread;
 use net_traits::image_cache_thread::{ImageCacheChan, ImageCacheThread};
-use net_traits::storage_thread::{StorageThread, StorageType};
 use num_traits::ToPrimitive;
 use page::Page;
 use profile_traits::mem;
@@ -72,6 +71,7 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::mpsc::TryRecvError::{Disconnected, Empty};
 use std::sync::mpsc::{Sender, channel};
 use std::sync::{Arc, Mutex};
+use storage_traits::storage_thread::{StorageThread, StorageType};
 use string_cache::Atom;
 use style::context::ReflowGoal;
 use style::error_reporting::ParseErrorReporter;
