+++ components/servo/lib.rs
@@ -39,6 +39,8 @@ pub extern crate profile;
 pub extern crate profile_traits;
 pub extern crate script;
 pub extern crate script_traits;
+pub extern crate storage;
+pub extern crate storage_traits;
 pub extern crate style;
 pub extern crate url;
 pub extern crate util;
@@ -73,14 +75,14 @@ use gfx::font_cache_thread::FontCacheThread;
 use ipc_channel::ipc::{self, IpcSender};
 use net::image_cache_thread::new_image_cache_thread;
 use net::resource_thread::new_resource_thread;
-use net::storage_thread::StorageThreadFactory;
-use net_traits::storage_thread::StorageThread;
 use profile::mem as profile_mem;
 use profile::time as profile_time;
 use profile_traits::mem;
 use profile_traits::time;
 use std::rc::Rc;
 use std::sync::mpsc::Sender;
+use storage::storage_thread::StorageThreadFactory;
+use storage_traits::storage_thread::StorageThread;
 use util::resource_files::resources_dir_path;
 use util::{opts, prefs};
