+++ components/storage/lib.rs
@@ -1,0 +1,10 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+extern crate ipc_channel;
+extern crate storage_traits;
+extern crate url;
+extern crate util;
+
+pub mod storage_thread;
