+++ components/storage/storage_thread.rs
@@ -3,14 +3,14 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use ipc_channel::ipc::{self, IpcReceiver, IpcSender};
-use net_traits::storage_thread::{StorageThread, StorageThreadMsg, StorageType};
-use resource_thread;
 use std::borrow::ToOwned;
 use std::collections::BTreeMap;
 use std::collections::HashMap;
+use storage_traits::storage_thread::{StorageThread, StorageThreadMsg, StorageType};
 use url::Url;
 use util::opts;
 use util::thread::spawn_named;
+use util::write::write_json_to_file;
 
 const QUOTA_SIZE_LIMIT: usize = 5 * 1024 * 1024;
 
@@ -72,7 +72,7 @@ impl StorageManager {
                 }
                 StorageThreadMsg::Exit => {
                     if let Some(ref profile_dir) = opts::get().profile_dir {
-                        resource_thread::write_json_to_file(&self.local_data, profile_dir, "local_data.json");
+                        write_json_to_file(&self.local_data, profile_dir, "local_data.json");
                     }
                     break
                 }
