+++ components/storage_traits/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "storage_traits"
+version = "0.1.0"
+publish = false
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+heapsize = "0.3.0"
+heapsize_plugin = "0.1.2"
+serde = "0.7"
+serde_macros = "0.7"
+url = {version = "0.5.7", features = ["heap_size"]}
+
+[dependencies.ipc-channel]
+git = "https://github.com/servo/ipc-channel"
