+++ components/util/lib.rs
@@ -71,6 +71,7 @@ pub mod time;
 pub mod vec;
 #[allow(unsafe_code)]
 pub mod workqueue;
+pub mod write;
 
 #[allow(unsafe_code)]
 pub fn breakpoint() {
